refactor

+++ src/main/java/com/battleship/enumeration/ShipType.java
@@ -5,7 +5,8 @@ public enum ShipType {
     SINGLE_DECKER(1), DOUBLE_DECKER(2), THREE_DECKER(3), FOUR_DECKER(4);
 
     int deckerCount;
-    double health;
+
+    int health;
 
     ShipType(int deckerCount) {
         this.deckerCount = deckerCount;
@@ -15,11 +16,11 @@ public enum ShipType {
         return deckerCount;
     }
 
-    public double getHealth() {
+    public int getHealth() {
         return health;
     }
 
-    public void setHealth(double health) {
+    public void setHealth(int health) {
         this.health = health;
     }
 }
